pyproject.toml
[project]
name = "banner_flip_engine_2"
version = "0.2.0"
description = "Identify and research banner flip opportunities."
authors = [
    { name = "Lewis Tribble" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.55",
    "langchain-community>=0.3.9",
    "langchain-openai>=0.3.7",
    "langchain-anthropic>=0.3.9",
    "openai>=1.61.0",
    "tavily-python>=0.5.0",
    "langchain-groq>=0.2.4",
    "exa-py>=1.8.9",
    "arxiv>=2.1.3",
    "pymupdf>=1.25.3",
    "xmltodict>=0.14.2",
    "linkup-sdk>=0.2.3",
    "duckduckgo-search>=3.0.0",
    "ipython>=8.18.1",
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "geopy>=2.4.1",
    "chainlit>=2.3.0",
    "beautifulsoup4>=4.13.3",
    "websockets>=15.0.1",
]
